add support for field id
Closes #20

This commit adds `id` option on `parquet` tag which is in the form of `id={n}`
where n is an `int32` denoting a field it.

The underlying infrastructure was already there. We just add api for users to
define and access field id to the parquet schema.

Example

```go
type MySchema struct {
	A string `parquet:"a,id=2"`
}
```

Will set column `a` field id to `2`.

Printing
-----

I wasn't so sure how to show the field id in the printed schema. I settled with
showing the id next to the field name. I am open to any suggestions on how to
improve field id printing.

The above schema is printed as

```
message {
	required binary a = 2 (STRING);
}
```